pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wiggle_light"
version = "0.0.1"
authors = [
  { name="Studio Rabota" },
]
description = "WiggleLight - Control the light of WiggleBin via CLI"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  'adafruit-circuitpython-neopixel-spi',
]

[project.scripts]
wiggle-light = "wiggle_light.main:main"

[project.urls]
"Homepage" = "https://github.com/wiggle-bin/wiggle-light.git"